Apple System Fonts never loaded.
There was a logic error in determining if a font was a system font.  We tested to see if the file name did not end in 'ttf' so we assumed it was a system font, however, in the skin loader function where the configs are read in it specifically checks to see if the file name ends in 'ttf', if it does not then the font is never loaded -- causing our Apple system font loader not to work.

Furthermore, we added logic to look in /System/Library/Fonts then /Library/Fonts for system fonts (only if the font was not already found in another location).