Fix #222: No hebrew subtitles
This was more of a side effect than anything else.  In an attempt to verify
that the Hebrew fonts would work using the latest revisions of the font
loading mechanisms introduced in commit 482a45, I found out that the fonts
in the settings menu were being incorrectly displayed.  They should display
available family fonts not all available fonts.

Furthermore, when a font was being loaded it would try to load a file name
based on the font's family display name not the actual file path.