New #87: [Feature] Always on top while in windowed mode
There is a new menu item called "Float on Top" under the View
menu.  Select this menu item to make the window float on top
press it again to un-toggle it.

If you wish to use this feature, you must enable it each time
you run Plex.  There is no associated setting for this feature
and the toggle state of this feature does not persist between
Plex instantiations.