fixed: WMP has an annoying habit of urldecoding upnp paths before requesting from server
The idea here is to prefix the path with %/ and then urlencoding it again. This way when we get it back, if prefix equals %25 everything is normal, but if it's %/ again, the server has invalidly decoded the path once. But since we do a "safety" encode, our path should be intact.

git-svn-id: http://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@30663 568bbfeb-2a22-0410-94d2-cc84cf5bfa90