fixed: getting codec name from external libavcodec
CDVDDemuxFFmpeg::GetStreamCodecName() calls m_dllAvCodec.IsLoaded()
before calling avcodec_find_decoder(). However, this breaks when using
external ffmpeg, as it is implemented as a fake DllDynamic and therefore
always returns false for IsLoaded().

Fix that by simply removing the check. It is unnecessary as m_streams[N]
are always NULL when m_dllAvCodec is not loaded, so the block is not
reached at all.

git-svn-id: http://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@31495 568bbfeb-2a22-0410-94d2-cc84cf5bfa90