added: strings for episode count postfixes
Add strings "episode" and "episodes" for use in "1 episode" and
"3 episodes", respectively, and use them as appropriate instead of
"Episodes".

While this doesn't make much of a difference in English, this makes a
huge difference on some other languages, such as Finnish, where there
are more differences between these use cases than just the
capitalization.

For the record, technically "Episodes: 3" use case should also be
separated, but there it doesn't matter as much (at least in Finnish), so
that is not done for now.

git-svn-id: http://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@31710 568bbfeb-2a22-0410-94d2-cc84cf5bfa90