fixed: missing LDFLAGS/CFLAGS when removed from environment
Ideally, Makefile.include.in should contain "CFLAGS =
user-specified-cflags" instead of "CFLAGS += user-specified-cflags",
and the individual Makefiles would use '+=' as needed.

This would allow the 'user-specified-cflags' from 'configure' to be
completely overridden (instead of just appended) with
"make CFLAGS=some-other-flags" as is the general convention.

Since that would be fairly disruptive, lets not do it for now.

git-svn-id: http://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@31737 568bbfeb-2a22-0410-94d2-cc84cf5bfa90