Prevent Plex directories from being passed on to the DVDPlayer
If a plugin returns a VideoItem that's resolved to a real video
file at playback the plugin has no way of reporting errors back
using eg. MessageContainer, as the video-item will be passed on
to the Plex playback machinery regardless of content-type.

The same situation can arrise for library items with parts that
point to a .strm file, that in turn point to a plugin that's
responsible for returning a redirect to a Stream.LocalFile().

The problem is solved by hooking in to one of the main entry
points for playback, CGUIWindowVideoBase::OnPlayMedia(), and
probing the requested item for a possible Plex directory. If
such a directory is found a redirect-item is built and the
playback action is circumvented by starting a new OnClick()
on the redirect-item instead.

The use of a non-virtual overload of OnClick() is intentional,
as a virtual overload would be hidden by a single overriden
overload in a derived class such as CGUIWindowVideoBase.

There are still code-paths in Plex that will result in the
invalid (non-video) item being passed on to the DVDPlayer,
for example playing through the video-info overlay.