Add XBMC HTTP API for updating the progress bar dialog
Useful for allowing plugins which take a while to return something to
give feedback to the user (including progress), instead of the generic
"Waiting for response..." message.

An alterante way to achive the same thing would be through the use of
ExecBuiltIn(RunScript()) with a python-script that would use the xbmcgui
module to update the dialog, but XBMC python scripts have been disabled
in Plex.

This patch is a workaround, pending a similar feature being introduced
in the Plex Media Server plugin framework.